Update is done according to template.
