+++ test_scripts/Polices/Policy_Table_Update/001_ATF_P_TC_PTU_AppID_NotListed_PT_DeviceConsented.lua
@@ -1,0 +1,105 @@
+---------------------------------------------------------------------------------------------
+-- Requirements summary:
+-- [PolicyTableUpdate] Request PTU - an app registered is not listed in PT (device consented)
+--
+-- Description:
+-- SDL should request PTU in case new application is registered and is not listed in PT
+-- and device is consented.
+-- 1. Used preconditions
+-- SDL is built with "-DEXTENDED_POLICY: EXTERNAL_PROPRIETARY" flag
+-- Connect mobile phone over WiFi.
+-- Register new application.
+-- Successful PTU. Device is consented.
+-- 2. Performed steps
+-- Register new application
+--
+-- Expected result:
+-- PTU is requested. PTS is created.
+-- SDL->HMI: SDL.OnStatusUpdate(UPDATE_NEEDED)
+-- SDL->HMI: BasicCommunication.PolicyUpdate
+---------------------------------------------------------------------------------------------
+
+--[[ General configuration parameters ]]
+config.deviceMAC = "12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0"
+
+--[[ Required Shared libraries ]]
+local commonSteps = require('user_modules/shared_testcases/commonSteps')
+local commonFunctions = require('user_modules/shared_testcases/commonFunctions')
+local commonPreconditions = require ('user_modules/shared_testcases/commonPreconditions')
+local testCasesForBuildingSDLPolicyFlag = require('user_modules/shared_testcases/testCasesForBuildingSDLPolicyFlag')
+local testCasesForPolicyTableSnapshot = require('user_modules/shared_testcases/testCasesForPolicyTableSnapshot')
+
+--[[ General Precondition before ATF start ]]
+-- commonFunctions:SDLForceStop()
+testCasesForBuildingSDLPolicyFlag:Update_PolicyFlag("EXTENDED_POLICY", "EXTERNAL_PROPRIETARY")
+testCasesForBuildingSDLPolicyFlag:CheckPolicyFlagAfterBuild("EXTENDED_POLICY","EXTERNAL_PROPRIETARY")
+commonSteps:DeleteLogsFileAndPolicyTable()
+commonPreconditions:Connecttest_without_ExitBySDLDisconnect_WithoutOpenConnectionRegisterApp("connecttest_RAI.lua")
+
+--ToDo: shall be removed when issue: "ATF does not stop HB timers by closing session and connection" is fixed
+config.defaultProtocolVersion = 2
+
+--[[ General Settings for configuration ]]
+Test = require('user_modules/connecttest_RAI')
+require('cardinalities')
+require('user_modules/AppTypes')
+local mobile_session = require('mobile_session')
+
+--[[ Preconditions ]]
+commonFunctions:newTestCasesGroup("Preconditions")
+function Test:Precondition_remove_user_connecttest()
+  os.execute( "rm -f ./user_modules/connecttest_RAI.lua" )
+end
+
+--[[ Test ]]
+commonFunctions:newTestCasesGroup("Test")
+function Test:TestStep_ConnectMobile()
+  self:connectMobile()
+end
+
+function Test:TestStep_StartNewSession()
+  self.mobileSession = mobile_session.MobileSession( self, self.mobileConnection)
+  self.mobileSession:StartService(7)
+end
+
+function Test:TestStep_PTU_AppID_NotListed_PT()
+
+  local correlationId = self.mobileSession:SendRPC("RegisterAppInterface", config.application1.registerAppInterfaceParams)
+
+  EXPECT_HMINOTIFICATION("BasicCommunication.OnAppRegistered", { application = { appName = config.application1.appName } })
+  :Do(function(_,data)
+      local hmi_app_id = data.params.application.appID
+      EXPECT_HMINOTIFICATION("SDL.OnStatusUpdate", {status = "UPDATE_NEEDED"})
+
+      testCasesForPolicyTableSnapshot:create_PTS(true,
+        {config.application1.registerAppInterfaceParams.appID},
+        {config.deviceMAC},
+        {hmi_app_id})
+
+      local timeout_after_x_seconds = testCasesForPolicyTableSnapshot:get_data_from_PTS("module_config.timeout_after_x_seconds")
+      local seconds_between_retries = {}
+
+      for i = 1, #testCasesForPolicyTableSnapshot.pts_seconds_between_retries do
+        seconds_between_retries[i] = testCasesForPolicyTableSnapshot.pts_seconds_between_retries[i].value
+      end
+
+      EXPECT_HMICALL("BasicCommunication.PolicyUpdate",
+        {
+          file = "/tmp/fs/mp/images/ivsu_cache/PolicyTableUpdate",
+          timeout = timeout_after_x_seconds,
+          retry = seconds_between_retries
+        })
+      :Do(function(_,data)
+          self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS", {})
+        end)
+    end)
+  self.mobileSession:ExpectResponse(correlationId, { success = true, resultCode = "SUCCESS"})
+end
+
+--[[ Postconditions ]]
+commonFunctions:newTestCasesGroup("Postconditions")
+function Test:Postcondition_SDLForceStop()
+  commonFunctions:SDLForceStop(self)
+end
+
+return Test
